+++ src/skills/Entries.php
@@ -64,9 +64,9 @@ class Entries extends BaseSkillSet
 
         // If no results
         if (!$results) {
-            // Return error message
+            // Return success message with no results
             return new SkillResponse([
-                'success' => false,
+                'success' => true,
                 'message' => "No entries found{$inSection}."
             ]);
         }