+++ src/skills/Tags.php
@@ -89,9 +89,9 @@ class Tags extends BaseSkillSet
 
         // If no results
         if (!$results) {
-            // Return error message
+            // Return success message with no results
             return new SkillResponse([
-                'success' => false,
+                'success' => true,
                 'message' => "No tags found{$inGroup}."
             ]);
         }